cleaning up the assets traits
